src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.6
idna==3.2
PyYAML==5.4.1
requests==2.26.0
urllib3==1.26.7
